split analysis results to 2 tabs
