+++ windwatts-ui/src/components/resultPane/details/DetailsTab.tsx
@@ -1,0 +1,48 @@
+import { Stack } from "@mui/material";
+import { getOutOfBoundsMessage } from "../../../utils";
+import { useOutputUnit, useProductionData } from "../../../hooks";
+import { ProductionDataTable } from "./ProductionDataTable";
+import { WindRose } from "./WindRose";
+
+export const DetailsTab = () => {
+  const {
+    productionData,
+    isLoading: productionLoading,
+    error: productionError,
+    hasData: productionHasData,
+    outOfBounds,
+    dataModel,
+    lat,
+    lng,
+  } = useProductionData();
+
+  useOutputUnit(); // auto-switches between kWh and MWh
+
+  const hasMonthlyData =
+    productionData && "monthly_avg_energy_production" in productionData;
+  const tableData = hasMonthlyData
+    ? productionData?.monthly_avg_energy_production
+    : productionData?.yearly_avg_energy_production;
+  const timeUnit = hasMonthlyData ? "month" : "year";
+
+  const outOfBoundsMessage = outOfBounds
+    ? getOutOfBoundsMessage(lat, lng, dataModel)
+    : undefined;
+
+  return (
+    <Stack spacing={2}>
+      <WindRose />
+
+      <ProductionDataTable
+        title=""
+        data={tableData}
+        timeUnit={timeUnit}
+        isLoading={productionLoading}
+        hasData={productionHasData}
+        error={productionError}
+        outOfBoundsMessage={outOfBoundsMessage}
+        emptyMessage="Select a location and turbine to see production details."
+      />
+    </Stack>
+  );
+};
