+++ windwatts-ui/src/components/resultPane/details/index.ts
@@ -1,0 +1,3 @@
+export { ProductionDataTable } from "./ProductionDataTable";
+export { WindRose } from "./WindRose";
+export { DetailsTab } from "./DetailsTab";
